+++ Source/Workbench/Features/EventStore/Namespaces/Observers/ObserverRunningStateHelpers.tsx
@@ -4,38 +4,7 @@
 import { ObserverRunningState } from 'Api/Concepts/Observation/ObserverRunningState';
 import { ColumnFilterElementTemplateOptions } from 'primereact/column';
 import { MultiSelect, MultiSelectChangeEvent } from 'primereact/multiselect';
-
-export const getObserverRunningStateAsText = (
-    runningState: ObserverRunningState | string
-) => {
-    switch (runningState) {
-        case ObserverRunningState.new:
-            return 'New';
-        case ObserverRunningState.routing:
-            return 'Routing';
-        case ObserverRunningState.replaying:
-            return 'Replaying';
-        case ObserverRunningState.catchingUp:
-            return 'CatchingUp';
-        case ObserverRunningState.active:
-            return 'Active';
-        case ObserverRunningState.paused:
-            return 'Paused';
-        case ObserverRunningState.stopped:
-            return 'Stopped';
-        case ObserverRunningState.suspended:
-            return 'Suspended';
-        case ObserverRunningState.failed:
-            return 'Failed';
-        case ObserverRunningState.tailOfReplay:
-            return 'TailOfReplay';
-        case ObserverRunningState.disconnected:
-            return 'Disconnected';
-        case ObserverRunningState.indexing:
-            return 'Indexing';
-    }
-    return '[N/A]';
-};
+import { getObserverRunningStateAsText } from './getObserverRunningStateAsText';
 
 interface ObserverRunningStateOption {
     name: string;
